+++ packages/dds/tree/src/index.ts
@@ -371,4 +371,4 @@ export { TableSchema, type System_TableSchema } from "./tableSchema.js";
 export { asAlpha, asBeta } from "./api.js";
 
 export { TextAsTree, FormattedTextAsTree } from "./text/index.js";
-export { ExtensibleUnionNode } from "./extensibleSchemaUnion.js";
+export { ExtensibleUnionNode } from "./extensibleUnionNode.js";
